+++ gulpfile.js
@@ -202,7 +202,7 @@ function makePackage(platform, arch, callback) {
     icon: 'resources/icon',
     "version-string": {
       CompanyName: distPackageAuthor,
-      LegalCopyright: `Copyright (c) 2015 - 2016 ${packageJson.author.name}`,
+      LegalCopyright: `Copyright (c) 2015 - ${new Date().getFullYear()} ${packageJson.author.name}`,
       FileDescription: packageJson.description,
       OriginalFilename: packageJson.productName + '.exe',
       ProductVersion: packageJson.version,