+++ src/main.js
@@ -11,12 +11,27 @@ const {
   systemPreferences
 } = require('electron');
 
-if (require('electron-squirrel-startup')) app.quit();
+if (process.platform === 'win32') {
+  var cmd = process.argv[1];
+  if (cmd === '--squirrel-uninstall') {
+    var AutoLaunch = require('auto-launch');
+    var appLauncher = new AutoLaunch({
+      name: 'Mattermost'
+    });
+    appLauncher.isEnabled().then(function(enabled) {
+      if (enabled)
+        appLauncher.disable();
+    });
+  }
+}
+
+require('electron-squirrel-startup');
 
 const fs = require('fs');
 const path = require('path');
 
 var settings = require('./common/settings');
+const osVersion = require('./common/osVersion');
 var certificateStore = require('./main/certificateStore').load(path.resolve(app.getPath('userData'), 'certificate.json'));
 var appMenu = require('./main/menus/app');
 const allowProtocolDialog = require('./main/allowProtocolDialog');
@@ -218,11 +233,20 @@ app.on('ready', function() {
       mainWindow.focus();
     });
     ipcMain.on('notified', function(event, arg) {
-      trayIcon.displayBalloon({
-        icon: path.resolve(__dirname, 'resources/appicon.png'),
-        title: arg.title,
-        content: arg.options.body
-      });
+      if (process.platform === 'win32') {
+        if (config.notifications.flashWindow === 2) {
+          mainWindow.flashFrame(true);
+        }
+        // On Windows 8.1 and Windows 8, a shortcut with a Application User Model ID must be installed to the Start screen.
+        // In current version, use tray balloon for notification
+        if (osVersion.isLowerThanOrEqualWindows8_1()) {
+          trayIcon.displayBalloon({
+            icon: path.resolve(__dirname, 'resources/appicon.png'),
+            title: arg.title,
+            content: arg.options.body
+          });
+        }
+      }
     });
 
     // Set overlay icon from dataURL